remove dynamic import
