+++ src/deps/webRepositoryAdapter.js
@@ -1,4 +1,4 @@
-// Repository Storage Adapters
+// Web Repository Storage Adapter
 
 /**
  * IndexedDB Repository Adapter for web environment
@@ -52,48 +52,6 @@ export const createIndexeddbRepositoryAdapter = () => {
         request.onsuccess = () => resolve(request.result || []);
         request.onerror = () => reject(request.error);
       });
-    },
-  };
-};
-
-/**
- * Tauri SQLite Repository Adapter
- */
-export const createTauriSQLiteRepositoryAdapter = async () => {
-  // Dynamic import to avoid loading in non-Tauri environments
-  const Database = (await import("@tauri-apps/plugin-sql")).default;
-
-  // Initialize SQLite database
-  const db = await Database.load("sqlite:repository.db");
-
-  // Create actions table if it doesn't exist
-  await db.execute(`
-    CREATE TABLE IF NOT EXISTS actions (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      action_type TEXT NOT NULL,
-      target TEXT,
-      value TEXT,
-      created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-    )
-  `);
-
-  return {
-    async addAction(action) {
-      await db.execute(
-        "INSERT INTO actions (action_type, target, value) VALUES ($1, $2, $3)",
-        [action.actionType, action.target, JSON.stringify(action.value)],
-      );
-    },
-
-    async getAllEvents() {
-      const results = await db.select(
-        "SELECT action_type, target, value FROM actions ORDER BY id",
-      );
-      return results.map((row) => ({
-        actionType: row.action_type,
-        target: row.target,
-        value: row.value ? JSON.parse(row.value) : null,
-      }));
     },
   };
 };
